Updated report and packages
