+++ src/meetix/lib/view/conferences/ActiveConferencesPage.dart
@@ -5,27 +5,26 @@ import 'package:meetix/controller/FirestoreController.dart';
 import 'package:meetix/controller/FunctionsController.dart';
 import 'package:meetix/controller/StorageController.dart';
 import 'package:meetix/model/Conference.dart';
-import 'package:meetix/view/ConferencePage.dart';
-import 'package:meetix/view/CreateProfilePage.dart';
-import 'package:meetix/view/MyWidgets.dart';
+import 'package:meetix/view/conferences/ConferencePage.dart';
+import 'package:meetix/view/profiles/CreateProfilePage.dart';
+import 'package:meetix/view/meetix_widgets/MyWidgets.dart';
 import 'package:provider/provider.dart';
 
-
-class MyCreatedConferencesPage extends StatefulWidget {
+class ActiveConferencesPage extends StatefulWidget {
   final FirestoreController _firestore;
   final StorageController _storage;
   final FunctionsController _functions;
   @required final Function(int) onChangeConfTab;
 
-  MyCreatedConferencesPage(this._firestore, this._storage, this._functions, {this.onChangeConfTab});
+  ActiveConferencesPage(this._firestore, this._storage, this._functions, {this.onChangeConfTab});
 
   @override
-  _MyCreatedConferencesPageState createState() {
-    return _MyCreatedConferencesPageState();
+  _ActiveConferencesPageState createState() {
+    return _ActiveConferencesPageState();
   }
 }
 
-class _MyCreatedConferencesPageState extends State<MyCreatedConferencesPage> {
+class _ActiveConferencesPageState extends State<ActiveConferencesPage> {
   @override
   Widget build(BuildContext context) {
     return _buildBody(context);
@@ -34,18 +33,20 @@ class _MyCreatedConferencesPageState extends State<MyCreatedConferencesPage> {
   Widget _buildBody(BuildContext context) {
     // Gets stream from Firestore with the conference info
     return StreamBuilder<QuerySnapshot>(
-      stream: widget._firestore.getCreatedConferences(context.watch<AuthController>().currentUser.uid),
+      stream: widget._firestore.getActiveConferences(),
       builder: (context, snapshot) {
         if (snapshot.hasData) {
           if (snapshot.data.size > 0) {
             return _buildList(context, snapshot.data.docs);
           } else {
-            return Center(child: Text("You have not created any conference"));
+            return Center(child: Text("There are no active conferences"));
           }
         } else if (snapshot.hasError) {
           return Text("Error :(");
         } else {
-          return Center(child: CircularProgressIndicator());
+          return Center(child: Text("Loading"));
+
+          // return Center(child: CircularProgressIndicator());
         }
       },
     );
@@ -66,7 +67,7 @@ class _MyCreatedConferencesPageState extends State<MyCreatedConferencesPage> {
 
     return InkWell(
       onTap: () {
-        Navigator.push(context, MaterialPageRoute(builder: (context) => openConference(context, _conference))).then((value) => setState((){}));
+        Navigator.push(context, MaterialPageRoute(builder: (context) => openConference(context, _conference))).then((value) => setState((){print("UPDATING");}));
       },
       child: Padding(
         padding: const EdgeInsets.all(16.0),
