+++ src/meetix/lib/view/conferences/EditConferencePage.dart
@@ -6,7 +6,7 @@ import 'package:meetix/controller/FirestoreController.dart';
 import 'package:meetix/controller/FunctionsController.dart';
 import 'package:meetix/controller/StorageController.dart';
 import 'package:meetix/model/Conference.dart';
-import 'package:meetix/view/MyWidgets.dart';
+import 'package:meetix/view/meetix_widgets/MyWidgets.dart';
 
 
 class EditConferencePage extends StatefulWidget {
