+++ src/meetix/lib/view/conferences/MyJoinedConferencesPage.dart
@@ -5,8 +5,8 @@ import 'package:meetix/controller/FirestoreController.dart';
 import 'package:meetix/controller/FunctionsController.dart';
 import 'package:meetix/controller/StorageController.dart';
 import 'package:meetix/model/Conference.dart';
-import 'package:meetix/view/ConferencePage.dart';
-import 'package:meetix/view/MyWidgets.dart';
+import 'package:meetix/view/conferences/ConferencePage.dart';
+import 'package:meetix/view/meetix_widgets/MyWidgets.dart';
 import 'package:provider/provider.dart';
