+++ src/meetix/lib/view/conferences/TopProfilesPage.dart
@@ -6,9 +6,9 @@ import 'package:meetix/controller/FunctionsController.dart';
 import 'package:meetix/controller/StorageController.dart';
 import 'package:meetix/model/Conference.dart';
 import 'package:meetix/model/Profile.dart';
-import 'package:meetix/view/ViewProfileDetailsPage.dart';
+import 'package:meetix/view/profiles/ViewProfileDetailsPage.dart';
 import 'package:provider/provider.dart';
-import 'MyWidgets.dart';
+import '../meetix_widgets/MyWidgets.dart';
 
 class TopProfilesPage extends StatefulWidget {
   final FirestoreController _firestore;
