+++ src/meetix/lib/view/profiles/ViewProfileDetailsPage.dart
@@ -5,8 +5,8 @@ import 'package:meetix/controller/FirestoreController.dart';
 import 'package:meetix/model/Conference.dart';
 import 'package:meetix/model/Profile.dart';
 import 'package:meetix/controller/StorageController.dart';
-import 'package:meetix/view/EditProfilePage.dart';
-import 'package:meetix/view/MyWidgets.dart';
+import 'package:meetix/view/profiles/EditProfilePage.dart';
+import 'package:meetix/view/meetix_widgets/MyWidgets.dart';
 import 'package:provider/provider.dart';
 
 class ViewProfileDetailsPage extends StatefulWidget {
